backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
langchain>=0.0.340
openai>=1.3.0
coqui-tts>=0.0.13
python-multipart>=0.0.6
celery>=5.3.0
redis>=4.5.5
python-dotenv>=1.0.0
manim>=0.18.0
pdfplumber>=0.9.0
python-docx>=0.8.11
bpy>=3.4.0
pdfplumber>=0.9.0
python-docx>=0.8.11
pytesseract>=0.3.10
pillow>=10.0.0
manim-voiceover>=0.4.0
colour>=0.1.5
python-multipart>=0.0.6